+++ packages/extension-polkagate/src/assets/logos/index.ts
@@ -1,9 +1,9 @@
 // Copyright 2019-2024 @polkadot/extension-polkagate authors & contributors
 // SPDX-License-Identifier: Apache-2.0
-// @ts-nocheck
 
-export { default as logoWhite } from './pg-w.svg';
-export { default as logoPink } from './pg-p.svg';
-export { default as logoBlack } from './pg-b.svg';
-export { default as logoMotionLight } from './logoMotionLight.gif';
+//@ts-nocheck
 export { default as logoMotionDark } from './logoMotionDark.gif';
+export { default as logoMotionLight } from './logoMotionLight.gif';
+export { default as logoBlack } from './pg-b.svg';
+export { default as logoPink } from './pg-p.svg';
+export { default as logoWhite } from './pg-w.svg';