+++ packages/extension-polkagate/src/components/FullscreenChainNames.tsx
@@ -1,21 +1,20 @@
 // Copyright 2019-2024 @polkadot/extension-polkagate authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-// @ts-nocheck
 /* eslint-disable react/jsx-max-props-per-line */
 
 import type { DropdownOption } from '../util/types';
 
 import { FormControl, Grid, InputBase, MenuItem, Select, type SelectChangeEvent, type SxProps, type Theme, Typography } from '@mui/material';
 import { styled, useTheme } from '@mui/material/styles';
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import React, { useCallback, useContext, useEffect, useMemo, useState } from 'react';
 
-import { useGenesisHashOptions, useIsTestnetEnabled } from '@polkadot/extension-polkagate/src/hooks';
+import { useIsTestnetEnabled } from '@polkadot/extension-polkagate/src/hooks';
 import { TEST_NETS } from '@polkadot/extension-polkagate/src/util/constants';
 
-import getLogo from '../util/getLogo';
-import { sanitizeChainName, updateRecentChains } from '../util/utils';
+import { updateRecentChains } from '../util/utils';
 import ChainLogo from './ChainLogo';
+import { GenesisHashOptionsContext } from './contexts';
 import Label from './Label';
 
 interface Props {
@@ -52,15 +51,12 @@ const BootstrapInput = styled(InputBase)<{ address?: string | null }>(({ address
   }
 }));
 
-const Item: React.FC<{ height?: string, logoSize?: number, text: string }> = ({ height = '20px', logoSize = 19.8, text }) => {
-  const chainName = useCallback((text: string) => sanitizeChainName(text)?.toLowerCase(), []);
-  const logo = getLogo(chainName(text));
-
+const Item: React.FC<{ height?: string, logoSize?: number, text: string | undefined, genesisHash?: string }> = ({ genesisHash, height = '20px', logoSize = 19.8, text }) => {
   return (
     <Grid container height={height} justifyContent='flex-start'>
-      {text !== 'Allow use on any chain' && logo &&
+      {text !== 'Allow use on any chain' &&
         <Grid alignItems='center' container item pr='10px' width='fit-content'>
-          <ChainLogo chainName={text} size={logoSize} />
+          <ChainLogo chainName={text} genesisHash={genesisHash} size={logoSize} />
         </Grid>
       }
       <Grid alignItems='center' container item justifyContent='flex-start' width='fit-content'>
@@ -72,9 +68,10 @@ const Item: React.FC<{ height?: string, logoSize?: number, text: string }> = ({ 
   );
 };
 
-function FullscreenChain({ address, defaultValue, disabledItems, helperText, label, labelFontSize = '14px', onChange, options, style }: Props) {
+function FullscreenChainNames ({ address, defaultValue, disabledItems, helperText, label, labelFontSize = '14px', onChange, options, style }: Props) {
   const theme = useTheme();
-  const _allOptions = useGenesisHashOptions();
+  const _allOptions = useContext(GenesisHashOptionsContext);
+
   const isTestnetEnabled = useIsTestnetEnabled();
 
   const _options = useMemo(() => {
@@ -90,7 +87,7 @@ function FullscreenChain({ address, defaultValue, disabledItems, helperText, lab
     !isTestnetEnabled
       ? [...(disabledItems || []), ...TEST_NETS]
       : disabledItems
-    , [disabledItems, isTestnetEnabled]);
+  , [disabledItems, isTestnetEnabled]);
 
   useEffect(() => {
     onChange(defaultValue);
@@ -167,7 +164,7 @@ function FullscreenChain({ address, defaultValue, disabledItems, helperText, lab
                 const text = _options.find((option) => value === option.value || value === option.text)?.text?.split(/\s*\(/)[0];
 
                 return (
-                  <Item height='50px' logoSize={29} text={text} />
+                  <Item genesisHash={selectedValue} height='50px' logoSize={29} text={text} />
                 );
               }}
               sx={{
@@ -218,4 +215,4 @@ function FullscreenChain({ address, defaultValue, disabledItems, helperText, lab
   );
 }
 
-export default React.memo(FullscreenChain);
+export default React.memo(FullscreenChainNames);