+++ packages/extension-polkagate/src/fullscreen/homeFullScreen/index.tsx
@@ -17,7 +17,7 @@ import ProfileTabsFullScreen from './partials/ProfileTabsFullScreen';
 import TotalBalancePieChart, { type AssetsWithUiAndPrice } from './partials/TotalBalancePieChart';
 import WatchList from './partials/WatchList';
 
-export default function HomePageFullScreen (): React.ReactElement {
+function HomePageFullScreen (): React.ReactElement {
   useFullscreen();
   const { t } = useTranslation();
 
@@ -92,3 +92,5 @@ export default function HomePageFullScreen (): React.ReactElement {
     </Grid>
   );
 }
+
+export default React.memo(HomePageFullScreen);