+++ packages/extension-polkagate/src/popup/staking/pool/stake/createPool/index.tsx
@@ -59,7 +59,7 @@ export default function CreatePool(): React.ReactElement {
 
   const ED = api && api.consts['balances']['existentialDeposit'] as unknown as BN;
   const nextPoolId = poolStakingConsts && poolStakingConsts.lastPoolId.toNumber() + 1;
-  const DEFAULT_POOLNAME = `Polkagate 💜${nextPoolId ? ` - ${nextPoolId}` : ''}`;
+  const DEFAULT_POOLNAME = `PolkaGate 💜${nextPoolId ? ` - ${nextPoolId}` : ''}`;
   // const amountAsBN = useMemo(() => ED && (new BN(parseFloat(createAmount ?? '0') * 10 ** decimal)).sub(ED), [ED, createAmount, decimal]);
   const amountAsBN = useMemo(() => amountToMachine(createAmount, decimal), [createAmount, decimal]);